reconfigured SecurityConfig to use standard JDBC functionality for users and authoriries. Removed User entity and related classes. A Design is now related with a user by a database constraint with username as foreignkey instead of JPA mapping on a User entity
