+++ src/main/java/nl/novi/be_plantjesplanner/entities/Design.java
@@ -10,8 +10,8 @@ public class Design {
 
     private String title;
 
-    @OneToOne(mappedBy = "design")
-    private User user;
+    @Column(name = "username", nullable = false) //in the database this column contains the username as foreign key to the user table. The relation between a User and a Design is maintained by a database constraint on the Design table instead of a JPA mapping. see the file "schema.sql" for details  todo: dit is wel een design keuze. Ik doe het zo om op deze manier het beheren van users+ authorities helemaal over te laten aan JDBC.
+    private String username;
 
     public Long getId() {
         return id;
@@ -25,11 +25,11 @@ public class Design {
         this.title = title;
     }
 
-    public User getUser() {
-        return user;
+    public String getUsername() {
+        return username;
     }
 
-    public void setUser(User user) {
-        this.user = user;
-    }
+    public void setUsername(String username) {
+        this.username = username;
+    }//get setter voor username omdat het een foreign key van de user tabel is
 }
