+++ src/main/resources/data.sql
@@ -21,18 +21,20 @@ VALUES (1, 'klimroos.jpg', 'klimroos.jpg', '2025-04-28T20:45:00+02:00'),
        (2, 'zonnebloem.jpg', 'zonnebloem.jpg', '2025-04-28T20:45:00+02:00'),
        (3, 'Hosta-Plant.jpg', 'Hosta-Plant.jpg', '2025-04-28T20:45:00+02:00');
 
--- Insert Design
-INSERT INTO design (id, title) VALUES (1, 'Standaard ontwerp voor admin'),(2,'Mijn prachtige tuin');
+
 
 -- Insert Users (gebruik BCrypt hashes!) designer: "geheim" admin: "super_geheim"
-INSERT INTO users (id, username, password, enabled, role, creation_date, design_id) VALUES
-                                                                                        (1, 'designer_test', '$2a$12$SRLVtg4dIFlCKFpvM5mSveFVVM5wid0p8nL3I.omNHEPBJu2aRI8a', TRUE, 'DESIGNER', CURRENT_TIMESTAMP, 1),
-                                                                                        (2, 'admin_test',   '$2a$12$jKxKvO8s/EKuTiEghAQCPeXnMEizilOSKjb4.AXBjWZms8kRC20ry', TRUE, 'ADMIN', CURRENT_TIMESTAMP, 2);
+INSERT INTO users (username, password, enabled) VALUES
+                                                    ( 'designer_test', '$2a$12$SRLVtg4dIFlCKFpvM5mSveFVVM5wid0p8nL3I.omNHEPBJu2aRI8a', TRUE),
+                                                    ('admin_test',   '$2a$12$jKxKvO8s/EKuTiEghAQCPeXnMEizilOSKjb4.AXBjWZms8kRC20ry', TRUE);
 
 -- Authorities (rollen)
 INSERT INTO authorities (username, authority) VALUES
                                                   ('designer_test', 'ROLE_DESIGNER'),
                                                   ('admin_test',   'ROLE_ADMIN');
+-- Insert Design
+INSERT INTO design (id, title, username) VALUES (1, 'Standaard ontwerp voor admin','admin_test'),(2,'Mijn prachtige tuin','designer_test');
+
 
 -- Insert Plants
 INSERT INTO plants (id, dutch_name, latin_name, description, height, footprint, bloom_color_hex, bloom_color_group, published, locale_id, blooming_calendar_id, plantavatar_id)
@@ -41,10 +43,10 @@ VALUES
     (2, 'Zonnebloem', 'Helianthus annuus', 'Zonnige hoge bloem met groot hart.', 2.0, 0.5, '#FFD700', 'GEEL', true, 2, 2, 2),
     (3, 'Hosta', 'Hosta sieboldiana', 'Schaduwminnende vaste plant met brede bladeren.', 0.6, 0.8, '#98FB98', 'GROEN', true, 3, 3, 3);
 
--- 3. Sequences aanpassen (optioneel, afhankelijk van je ORM/sequences)
+-- 3. Sequences aanpassen
 SELECT setval('images_id_seq', (SELECT MAX(id) FROM images));
 SELECT setval('plants_id_seq', (SELECT MAX(id) FROM plants));
 SELECT setval('blooming_calendars_id_seq', (SELECT MAX(id) FROM blooming_calendars));
 SELECT setval('locales_id_seq', (SELECT MAX(id) FROM locales));
 SELECT setval('design_id_seq', (SELECT MAX(id) FROM design));
-SELECT setval('users_id_seq', (SELECT MAX(id) FROM users));
+-- SELECT setval('users_id_seq', (SELECT MAX(id) FROM users));
